+++ NEWS.md
@@ -1,5 +1,10 @@
 # hpa.spatial (development version)
 
+* default mb_geo in `map_data_with_correspondence()` as missing and apply 
+  default within function code. Using a hpa.spatial function as a default arg 
+  fails when it's called by namespace 
+  (`hpa.spatial::map_data_with_correspondence()`).
+
 # hpa.spatial 0.2.1
 
 * add `get_correspondence_tbl()` which calls `strayr::read_correspondence_tbl()`